requirements.txt
requests
python-dotenv
openai
python-aiml